requirements.txt
pandas>=2.0.0
numpy>=1.23.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
tensorflow-macos==2.12.0
tensorflow-metal==0.8.0
yfinance
